transaction service done

+++ src/main/java/org/postuniv/serviceAuto/repository/TransactionsRepository.java
@@ -9,9 +9,9 @@ import java.util.List;
 public interface TransactionsRepository {
     public boolean addTransaction(Transaction transaction);
     public Transaction getTransactionById(long transactionId);
-    public List<Transaction>  getTransactionByClientId(int clientId);
+    public List<Transaction>  getTransactionByClientId(long clientId);
     public List<Transaction> getTransactionByDate(LocalDateTime startingDate, LocalDateTime endingDate); //TODO: move to service.
-    public List<Transaction>  getTransactionByCarId(int carId);
+    public List<Transaction>  getTransactionByCarId(long carId);
     public List<Transaction> getAllTransactions();
     public boolean removeTransaction(long transactionId);
     public boolean updateTransaction(long transactionId,Transaction newTransaction);
